What I've learned here
For Cloudflare to cache the request that a worker makes using the
fetch function, the request needs to go to a hostname associated with
a Cloudflare zone and that needs to proxy to the origin. If the
worker sends a request to another worker route, the request will not
be cached. If the request is to a hostname not proxied at Cloudflare,
the request won't have a chance for caching.

The good news is that when I serve a cache-tag header from a site on
the platform, the request is clearable from the cache by that cache
tag.